tеst/DotComTest.java
import org.junit.Test;

import java.util.ArrayList;

import static org.junit.Assert.*;

/**
 * Created by stud_6 on 14.11.16.
 */
public class DotComTest {
    @Test
    public void checkYourself1() throws Exception {

        ArrayList<String> trueLocation = new ArrayList<String>();
        trueLocation.add("2");
        trueLocation.add("3");
        trueLocation.add("4");
        DotCom dotCom = new DotCom();
        String userGuess = "2";
        dotCom.setLocationCells(trueLocation);
        String result = dotCom.checkYourself(userGuess);
        String expResult = "Попал ";
        assertEquals(expResult, result);
    }


    @Test
    public void checkYourself2() throws Exception {

        ArrayList<String> trueLocation = new ArrayList<String>();
        trueLocation.add("2");
        trueLocation.add("3");
        trueLocation.add("4");
        DotCom dotCom = new DotCom();
        String userGuess = "1";
        dotCom.setLocationCells(trueLocation);
        String result = dotCom.checkYourself(userGuess);
        String expResult = "Мимо ";
        assertEquals(expResult, result);
    }
}





